+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### FEATURES:
 * builder/AmazonInstance: add support for packer's amazon instance builder
 * builder/Azure: add support for packer's azure builder
+* builder/Cloudstack: add support for packer's cloudstack builder
 * builder/ProfitBricks: add support for packer's ProfitBricks builder
 * builder/Qemu: add support for packer's qemu builder
 * builder/Triton: add support for packer's triton builder